requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20
deepface>=0.0.93
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tf-keras>=2.15.0
